+++ server/src/builtins/wavetablebuiltins.js
@@ -1371,6 +1371,17 @@ function createWavetableBuiltins() {
   );
 
   Builtin.createBuiltin(
+    "milliseconds-of",
+    ["len"],
+    function $millisecondsOf(env, executionEnvironment) {
+      let len = env.lb("len");
+      let ms = (convertTimeToSamples(len) / getSampleRate()) * 1000;
+      return constructFloat(ms);
+    },
+    "Returns the length of |len in milliseconds. |len takes a timebase tag like any other length, so this is how a length in beats becomes a number that something outside the audio system can use."
+  );
+
+  Builtin.createBuiltin(
     "get-bpm",
     [],
     function $getBpm(env, executionEnvironment) {
